PythonProject1/requirements.txt
speechrecognition
pyttsx3
requests
pywhatkit
spotipy
google-generative-ai
numpy
sounddevice
emoji
